requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
pynacl==1.5.0
google-generativeai==0.3.1
voicevox-client==0.2.0
ffmpeg-python==0.2.0
aiohttp==3.8.6
asyncio==3.4.3
SpeechRecognition==3.10.0
pyaudio==0.2.13
SQLAlchemy==2.0.23
aiosqlite==0.19.0
numpy==1.24.3
schedule==1.2.1
async-timeout==4.0.3
python-dateutil==2.8.2